requirements.txt
numpy==1.26.4
pillow==11.0.0
torch==2.2.2
torchvision==0.17.2
requests==2.32.3
matplotlib==3.9.2
scipy==1.14.1
transformers==4.46.2
huggingface-hub==0.26.2
gradio==5.5.0
timm==1.0.11
